Fixing win32 build

+++ src/gauge/time_benchmark.cpp
@@ -80,8 +80,10 @@ namespace gauge
 
         m_impl->m_stop = bc::high_resolution_clock::now();
         m_impl->m_stopped = true;
-        m_impl->m_result = bc::duration_cast<bc::microseconds>(
-            m_impl->m_stop-m_impl->m_start).count();
+
+        m_impl->m_result = static_cast<double>(
+            bc::duration_cast<bc::microseconds>(
+                m_impl->m_stop-m_impl->m_start).count());
 
         assert(m_impl->m_iterations > 0);
     }
@@ -105,7 +107,9 @@ namespace gauge
             if(factor > 2.0)
             {
                 // We seem to be running longer than needed
-                m_impl->m_iterations = (m_impl->m_iterations / factor) + 1;
+                m_impl->m_iterations = static_cast<uint32_t>(
+                    (m_impl->m_iterations / factor) + 1);
+
                 assert(m_impl->m_iterations > 0);
             }
 
@@ -134,7 +138,9 @@ namespace gauge
         assert(factor > 1.0);
 
         // Adjust the number of iterations with the factor
-        m_impl->m_iterations = (m_impl->m_iterations * factor) + 1;
+        m_impl->m_iterations = static_cast<uint32_t>(
+            (m_impl->m_iterations * factor) + 1);
+
         assert(m_impl->m_iterations > 0);
 
         return false;
